MG Color Palette

This function returns a character vector of color palettes used in the package.

This function returns a character vector of color palettes used in the package.

MG()

MG()

MG()

MG()

Value

A character vector of color hex codes.

A character vector of color hex codes.

A character vector of color hex codes.

Examples

# Example usage of the MG color palette
my_colors <- MG()
print(my_colors)
#>  [1] "#FFFF33"        "#FF7F00"        "#E41A1C"        "firebrick4"    
#>  [5] "#2e4057"        "#984EA3"        "#377EB8"        "olivedrab3"    
#>  [9] "#4DAF4A"        "#336633"        "grey80"         "#BB650B"       
#> [13] "gold"           "#559999"        "#7570b3"        "#E78AC3"       
#> [17] "#A6D854"        "#66a61e"        "#e6ab02"        "#a6761d"       
#> [21] "#663300"        "#66C2A5"        "#0e669b"        "#00798c"       
#> [25] "dodgerblue4"    "steelblue2"     "#00AFBB"        "#E7B800"       
#> [29] "#FC4E07"        "lightskyblue4"  "green"          "red"           
#> [33] "#FFF000"        "#0099CC"        "#FF9933"        "#CC9900"       
#> [37] "chartreuse1"    "#FF3399"        "#00FFFF"        "#0000CC"       
#> [41] "#A37F6F"        "#9183E6"        "#00AD9A"        "#990033"       
#> [45] "#909800"        "#00FF00"        "#17b5b4"        "#AED1D6"       
#> [49] "#b1010c"        "firebrick2"     "blue"           "navy"          
#> [53] "yellow"         "brown"          "black"          "purple"        
#> [57] "darkred"        "darkgreen"      "#82cfd0"        "#b2e0e4"       
#> [61] "honeydew3"      "#8d96a3"        "lavender"       "#CC6686"       
#> [65] "lavenderblush2" "mistyrose3"     "#e1deda"        "darkgoldenrod" 
#> [69] "burlywood"      "papayawhip"     "wheat4"         "cornsilk3"     
#> [73] "khaki2"         "beige"          "gray60"         "gray80"        
#> [77] "gray96"         "cadetblue4"     "honeydew2"      "mintcream"     
#> [81] "#0e668b"        "#a3c4dc"        "lightskyblue1"  "aliceblue"