Pre-process phyloseq object based on hashcodes

This function pre-processes a phyloseq object by subsetting, merging, and saving taxa based on specified hashcodes and a merge method (sum or max). It retains the taxa information and creates various intermediate and final datasets for further processing.

Pre_processing_hashcodes(
  spiked_16S,
  hashcodes,
  merge_method = c("sum", "max"),
  output_prefix = "merged_physeq"
)

Arguments

spiked_16S

A phyloseq object containing the spiked dataset.

hashcodes

A character vector of hashcodes (row names of the OTU table) to be processed.

merge_method

The method to use for merging taxa: "sum" or "max".

output_prefix

A character string to be used as a prefix for the output file names.

Value

A phyloseq object with the processed data.

Examples

if (FALSE) { # \dontrun{
Tetragenococcus <- phyloseq::subset_taxa(physeq_16SASV, 
  Species == "Tetragenococcus_halophilus" | Species == "Tetragenococcus_sp")
hashcodes <- row.names(phyloseq::otu_table(Tetragenococcus))
processed_data_sum <- Pre_processing_hashcodes(physeq_16SASV, hashcodes, 
  merge_method = "sum", output_prefix = "merged_physeq_sum")
processed_data_max <- Pre_processing_hashcodes(physeq_16SASV, hashcodes, 
  merge_method = "max", output_prefix = "merged_physeq_max")
summ_count_phyloseq(processed_data_sum)
} # }