Calculate Scaling Factors for Spiked Species in Phyloseq Object

This function calculates the scaling factors for specified spiked species in a phyloseq object. It excludes the spiked species, merges the spiked species into one ASV, calculates the scaling factors, and saves the results as RDS, CSV, and DOCX files.

calculate_spikeIn_factors(
  physeq,
  spiked_cells,
  merged_spiked_species,
  output_path = NULL
)

Arguments

physeq

A phyloseq object containing the microbial data.

spiked_cells

A numeric value specifying the number of spiked cells.

merged_spiked_species

A character vector of spiked species to merge in the phyloseq object.

output_path

A character string specifying the path to save the DOCX output file. Default is NULL, which saves the report as "spikeIn_factors_summary.docx".

Value

A list containing the scaling factors and other related outputs.

Details

Note: The metadata in the phyloseq object must contain a column named 'spiked.volume' with the appropriate values.

Examples

if (FALSE) { # \dontrun{
# Example usage:

# Define the spiked species
merged_spiked_species <- c("Tetragenococcus_halophilus")

# Calculate scaling factors and generate the report
result <- calculate_spikeIn_factors(merged_physeq_sum, 1874, merged_spiked_species)

# Access the results from the calculation
scaling_factors <- result$scaling_factors        # The calculated scaling factors
physeq_no_spiked <- result$physeq_no_spiked      # Phyloseq object with no spiked species
spiked_16S_total_reads <- result$spiked_16S_total_reads  # Total reads of the spiked 16S species
spiked_species <- result$spiked_species          # List of spiked species
spiked_species_merged <- result$spiked_species_merged  # Merged spiked species
spiked_species_reads <- result$spiked_species_reads    # Reads associated with the spiked species
} # }