Calculate Spike Percentage for Specified Taxa in a Phyloseq Object

This function calculates the percentage of reads from specified spiked species or hashcodes in a phyloseq object. It merges the spiked taxa into one ASV, calculates the percentage of reads, categorizes the results as passed or failed, and saves the results as a DOCX and CSV file.

calculate_spike_percentage(
  physeq,
  merged_spiked_species = NULL,
  merged_spiked_hashcodes = NULL,
  output_path = "merged_data.docx",
  passed_range = c(0.1, 11)
)

Arguments

physeq

A phyloseq object containing the microbial data.

merged_spiked_species

A character vector of spiked species to check in the phyloseq object. Default is NULL.

merged_spiked_hashcodes

A character vector of spiked hashcodes to check in the phyloseq object. Default is NULL.

output_path

A character string specifying the path to save the output files. Default is "merged_data.docx".

passed_range

A numeric vector of length 2 specifying the range of percentages to categorize results as "passed". Default is c(0.1, 11).

Value

A data frame containing the percentage of spiked taxa reads and the pass/fail results.

Examples

if (FALSE) { # \dontrun{
# Load example data
data(physeq_16SASV)
physeq <- physeq_16SASV

# Define the spiked species
merged_spiked_species <- c("Tetragenococcus_halophilus", "Tetragenococcus_sp.")

# Calculate spike percentage for the defined spiked species within the range 0.1 to 10
calculate_spike_percentage(physeq, 
                           merged_spiked_species = merged_spiked_species, 
                           passed_range = c(0.1, 10))

# Define the spiked hashcodes
merged_spiked_hashcodes <- c("hashcode1", "hashcode2")

# Calculate spike percentage for the defined spiked hashcodes within the range 0.1 to 10
calculate_spike_percentage(physeq,
                           merged_spiked_hashcodes = merged_spiked_hashcodes,
                           passed_range = c(0.1, 10))

# Example with a single species
merged_spiked_species <- "Dekkera_bruxellensis"

# Calculate spike percentage for the single species within the range 0.1 to 35
result <- calculate_spike_percentage(spiked_ITS_OTU_scaled, 
                                     merged_spiked_species, 
                                     passed_range = c(0.1, 35))

# Generate a summary statistics table from the result
calculate_summary_stats_table(result)

# Display the calculated spike percentage
result$Percentage
} # }