Calculate Spike Percentage for List of Specified Taxa in a Phyloseq Object

This function calculates the percentage of reads from specified spiked species in a phyloseq object. It merges the spiked taxa into one ASV, calculates the percentage of reads, categorizes the results as passed or failed, and saves the results as a DOCX and CSV file.

calculate_spike_percentage_list(
  physeq,
  merged_spiked_species = NULL,
  output_path = "merged_data.docx",
  passed_range = c(0.1, 11)
)

Arguments

physeq

A phyloseq object containing microbial data.

merged_spiked_species

A character vector or list of spiked species to check in the phyloseq object. Default is NULL.

output_path

A character string specifying the path to save the output files. Default is "merged_data.docx".

passed_range

A numeric vector of length 2 specifying the range of percentages to categorize results as "passed". Default is c(0.1, 11).

Value

A data frame containing the percentage of spiked taxa reads and the pass/fail results.

Examples

if (FALSE) { # \dontrun{
# Example usage:
spiked_species_list <- list(
  c("Pseudomonas aeruginosa"),
  c("Escherichia coli"),
  c("Clostridium difficile")
)

# Create a mock phyloseq object 
result <- calculate_spike_percentage_list(merged_physeq_sum, 
merged_spiked_species = spiked_species_list, 
passed_range = c(0.1, 10))

# Print the results
print(result)
} # }