Generate Spike Success Report and Summary Statistics

This function calculates the spike success report for a given phyloseq object, summarizes the results, and saves the summary statistics in both DOCX and CSV formats. It uses the `flextable` package to create and format the tables in the DOCX file.

conclusion(
  physeq,
  merged_spiked_species,
  max_passed_range = 11,
  output_path = NULL
)

Arguments

physeq

A phyloseq object containing the microbial data.

merged_spiked_species

A character vector of spiked species to check in the phyloseq object.

max_passed_range

A numeric value specifying the maximum acceptable spike percentage. Default is 11.

output_path

A character string specifying the path to save the output files. Default is NULL, which saves the report as "spike_success_report.docx".

Value

A data frame containing the summary statistics of the spike success report.

Examples

if (FALSE) { # \dontrun{
# Define the parameters
merged_spiked_species <- c("Tetragenococcus_halophilus") 
max_passed_range <- 12 
output_path <- "spike_success_report.docx"

# Convert the phyloseq object to absolute counts using scaling factors
absolute <- convert_to_absolute_counts(merged_physeq_sum, scaling_factors) 
absolute_counts <- absolute$absolute_counts 
physeq_absolute <- absolute$physeq_obj 

# Subset the phyloseq object to exclude blank samples (optional step)
physeq_16S_adj_scaled_perc <- phyloseq::subset_samples(physeq_absolute, sample.or.blank != "blank")

# Generate the spike success report and calculate summary statistics
summary_stats <- conclusion(physeq_16S_adj_scaled_perc, 
                            merged_spiked_species, 
                            max_passed_range, 
                            output_path)

# Print the summary statistics
print(summary_stats)
} # }