Perform Differential Abundance Analysis with DESeq2

Perform Differential Abundance Analysis with DESeq2

Perform and Visualize Differential Abundance Analysis This function normalizes and filters the data, performs differential abundance analysis, extracts significant OTUs, rebuilds the phyloseq object, merges results with metadata, and visualizes the results.

perform_and_visualize_differential_abundance(
  ps,
  group_var,
  method = "edgeR",
  threshold_percentage = 1e-04,
  threshold_mean_abundance = 1e-05,
  threshold_count = 1,
  threshold_relative_abundance = 1e-05,
  significance_level = 0.05,
  point_size = 3,
  facet_variable = "Phylum",
  target_glom = "Genus"
)

perform_and_visualize_differential_abundance(
  ps,
  group_var,
  method = "edgeR",
  threshold_percentage = 1e-04,
  threshold_mean_abundance = 1e-05,
  threshold_count = 1,
  threshold_relative_abundance = 1e-05,
  significance_level = 0.05,
  point_size = 3,
  facet_variable = "Phylum",
  target_glom = "Genus"
)

Arguments

ps

A phyloseq object containing the microbiome data.

group_var

A string specifying the grouping variable in sample data.

method

A string specifying the differential abundance method ("edgeR", "DESeq2"). Default is "edgeR".

threshold_percentage

A numeric value for filtering threshold percentage. Default is 0.0001.

threshold_mean_abundance

A numeric value for filtering mean abundance. Default is 0.00001.

threshold_count

A numeric value for filtering count. Default is 1.

threshold_relative_abundance

A numeric value for filtering relative abundance. Default is 0.00001.

significance_level

A numeric value specifying the significance level for filtering significant OTUs. Default is 0.05.

point_size

A numeric value specifying the size of points in the volcano plot. Default is 3.

facet_variable

A string specifying the variable to facet the barplots. Default is "Phylum".

target_glom

A string specifying the taxonomic rank to aggregate taxa (e.g., "Genus", "Family"). Default is "Genus".

Value

A data frame containing the differential abundance results.

A list containing the final results, the phyloseq object with significant OTUs, and ggplot objects for the visualizations.

A list containing the final results, the phyloseq object with significant OTUs, and ggplot objects for the visualizations.

Examples

if (FALSE) { # \dontrun{
# Install and load the required packages first
required_packages <- c("phyloseq", "DESeq2", "edgeR", 
                       "BiocManager", "BiocGenerics", "ggplot2", 
                       "dplyr", "DT")
install_and_load(required_packages)

# Step 1: Clean and preprocess the phyloseq object
ps <- remove_zero_negative_count_samples(ps)  
ps <- convert_categorical_to_factors(ps)      

# edgeR Example with default glomming at the "Genus" rank
results_edgeR <- perform_and_visualize_differential_abundance(ps, "Animal.ecomode", 
                                                              method = "edgeR", 
                                                              threshold_percentage = 0.001, 
                                                              threshold_mean_abundance = 0.001,
                                                              threshold_count = 5, 
                                                              threshold_relative_abundance = 0.001,
                                                              significance_level = 0.05, 
                                                              point_size = 3, 
                                                              target_glom = "Genus")

# Print the volcano plot and bar plots for edgeR results
print(results_edgeR$plot)
print(results_edgeR$barplot_rel)  # Relative abundance 
print(results_edgeR$barplot_abs)  # Absolute abundance 

# DESeq2 Example with custom glomming at the "Family" rank and faceting by "Class"
results_DESeq2 <- perform_and_visualize_differential_abundance(ps, "Host.genus", 
                                                               method = "DESeq2",
                                                               threshold_percentage = 0.001, 
                                                               threshold_mean_abundance = 0.001,
                                                               threshold_count = 5, 
                                                               threshold_relative_abundance = 0.001,
                                                               significance_level = 0.05, 
                                                               point_size = 3, 
                                                               target_glom = "Family", 
                                                               facet_variable = "Class")

# Print the volcano plot and bar plots for DESeq2 results
print(results_DESeq2$plot)
print(results_DESeq2$barplot_rel)  # Relative abundance 
print(results_DESeq2$barplot_abs)  # Absolute abundance 
} # }