Plot a Phylogenetic Tree with Glommed OTUs

This function gloms the OTUs in the phyloseq object at the specified resolution and plots the tree.

plot_glommed_tree(
  physeq,
  resolution,
  output_prefix = "glommed_tree",
  width = 20,
  height = 18,
  print_plot = TRUE
)

Arguments

physeq

A phyloseq object containing the phylogenetic data.

resolution

A numeric value specifying the resolution for glomming the OTUs.

output_prefix

A character string specifying the prefix for the output files. Default is "glommed_tree".

width

A numeric value specifying the width of the output plot. Default is 18.

height

A numeric value specifying the height of the output plot. Default is 18.

print_plot

A logical value specifying whether to print the plot. Default is TRUE.

Value

NULL. The function saves the plot to the specified output file.

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
  # Plot the tree with glommed OTUs
  plot_glommed_tree(Tetragenococcus, resolution = 0.2, 
  output_prefix = "top", width = 18, height = 18)
}
} # }