Plot Bar Abundance for Phyloseq Data

This function generates bar plots for phyloseq data at a specified taxonomic level, with options to customize the appearance, relativize the data or plot absolute abundance, facet the plots, and save the plots.

plotbar_abundance(
  physeq,
  level = "Genus",
  color = NULL,
  group = NULL,
  x_axis_var = NULL,
  top = 20,
  return = FALSE,
  x_size = 12,
  y_size = 12,
  legend_key_size = 1.5,
  legend_text_size = 12,
  legend_nrow = 20,
  relativize = TRUE,
  output_prefix = NULL,
  facet_var = NULL,
  scales = "fixed"
)

Arguments

physeq

A phyloseq object containing the microbiome data.

level

A character string specifying the taxonomic level to plot (e.g., "Genus", "Family").

color

A character vector specifying colors to use for the different taxa. Default is NULL.

group

A character vector specifying the grouping variables. Default is NULL.

x_axis_var

A character string specifying the variable(s) to be shown on the x-axis. Default is NULL.

top

An integer specifying the number of top taxa to display. Default is 20.

return

A logical indicating whether to return the summarized data frame. Default is FALSE.

x_size

An integer specifying the size of the x-axis text. Default is 12.

y_size

An integer specifying the size of the y-axis text. Default is 12.

legend_key_size

A numeric value specifying the size of the legend keys. Default is 1.5.

legend_text_size

An integer specifying the size of the legend text. Default is 12.

legend_nrow

An integer specifying the number of rows for the legend. Default is 20.

relativize

A logical indicating whether to relativize the data. Default is TRUE.

output_prefix

A character string specifying the prefix for the output file names. Default is NULL.

facet_var

A character vector specifying the faceting variables. Default is NULL.

scales

A character string specifying scales in facet. Options are "fixed", "free", "free_x", "free_y". Default is "fixed".

Value

A ggplot object if return is FALSE, otherwise a data frame with summarized data.

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
  # Example usage of plotbar_abundance
  plot <- plotbar_abundance(
    physeq_16SOTU,                   # phyloseq obj
    level = "Family",                 # Taxonomic level
    group = c("Diet", "Host.species", "Ecoregion.III"),  # Grouping variables
    x_axis_var = "Ecoregion.III",              # X-axis variable
    top = 10,                         # Number of top taxa 
    x_size = 10,                      # X-axis text size
    y_size = 10,                      # Y-axis text size
    legend_key_size = 0.5,            # Legend key size
    legend_text_size = 11,            # Legend text size
    legend_nrow = 10,                 # Legend Number of rows
    relativize = TRUE,                # Whether to show relative/absolute
    output_prefix = "rel_abundance_plot",  # File output prefix
    facet_var = "Diet",      # Facet by variable
    scales = "free_x"                 # Allow free scales for the x-axis/or y
  )
  print(plot) + ggtitle("Relative Abundance Across Ecoregions")
}
} # }