Subsampling to an Equal Sequencing Depth

This function performs subsampling to an equal sequencing depth, determined by the sample with the lowest sequencing depth after excluding very low abundant taxa. It rounds down the result to the nearest integer (floor). Caution: you might lose samples.

randomsubsample_Trimmed_evenDepth(
  physeq,
  smalltrim = 0.001,
  replace = TRUE,
  output_file = "randomsubsample_Trimmed_evenDepth.rds"
)

Arguments

physeq

A phyloseq object containing the microbial data.

smalltrim

A numeric value specifying the trimming percentage to exclude very low abundant taxa. Default is 0.001.

replace

A logical value indicating whether to sample with replacement. Default is TRUE.

output_file

A character string specifying the output file name for the subsampled phyloseq object. Default is "randomsubsample_Trimmed_evenDepth.rds".

Value

A phyloseq object with subsampled sequencing depths.

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
  # Perform subsampling
  spiked_ITS_evenDepth <- randomsubsample_Trimmed_evenDepth(spiked_ITS, smalltrim = 0.001)
}
} # }