Generate Ridge Plots to Visualize the Distribution of Abundances of Taxa

This function rarefies the data, performs proportion transformation, gloms and prunes taxa, and generates ridge plots to visualize the distribution of relative abundances of taxa.

ridge_plot_it(physeq, taxrank = "Genus", rarefaction_depth = NULL, top_n = 10)

Arguments

physeq

A phyloseq object containing the taxonomic and abundance data.

taxrank

A character string specifying the taxonomic rank to use for glomming and plotting. Default is "Genus".

rarefaction_depth

A numeric value specifying the rarefaction depth. If NULL, it is set to 90% of the minimum sample sum. Default is NULL.

top_n

An integer specifying the number of top taxa to include in the plot. Default is 10.

Value

A ggplot2 object representing the ridge plot of the distribution of relative abundances of taxa.

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
  ridgeP <- ridge_plot_it(spiked_16S, taxrank = "Family", top_n = 10)
  print(ridgeP) + my_custom_theme()
}
} # }