Create a Taxa Barplot

This function generates a bar plot of the relative or absolute abundances of taxa at a specified taxonomic rank. The top taxa are selected, and the plot can be customized with various options.

taxa_barplot(
  physeq,
  target_glom = "Genus",
  custom_tax_names = NULL,
  normalize = TRUE,
  treatment_variable = "Treatment",
  abundance_type = "relative",
  x_angle = 25,
  fill_variable = target_glom,
  facet_variable = "Phylum",
  top_n_taxa = 20,
  palette = MG()
)

Arguments

physeq

A phyloseq object containing the microbiome data.

target_glom

A character string specifying the taxonomic rank to plot (e.g., "Genus").

custom_tax_names

A character vector specifying custom taxonomic names for the levels. Default is NULL.

normalize

A logical indicating whether to normalize the sample counts to relative abundances. Default is TRUE.

treatment_variable

A character string specifying the treatment variable to use for the x-axis. Default is "Treatment".

abundance_type

A character string specifying whether to plot "relative" or "absolute" abundance. Default is "relative".

x_angle

A numeric value specifying the angle of the x-axis text labels. Default is 20.

fill_variable

A character string specifying the variable to use for fill in stacking taxa. Default is target_glom.

facet_variable

A character string specifying the variable to use for faceting. Default is "Phylum".

top_n_taxa

A numeric value specifying the number of top taxa to include in the plot. Default is 20.

palette

A character vector of color hex codes to use for the fill colors. Default is `MG()`.

Value

A list containing the ggplot2 bar plot object (`barplot`) and the pruned phyloseq object (`taxa_data`) with the top taxa.

Examples

if (FALSE) { # \dontrun{
# Generate a taxa barplot for the Genus rank with absolute abundance
bp_ab <- taxa_barplot(physeq_16SASV, 
                      target_glom = "Genus",
                      treatment_variable = "Host.species", 
                      abundance_type = "absolute", 
                      x_angle = 90,
                      fill_variable = "Genus", 
                      facet_variable = "Diet", 
                      top_n_taxa = 20, 
                      palette = MG())

# Print the barplot for absolute abundance
print(bp_ab$barplot)

# Generate a taxa barplot for the Genus rank with relative abundance
bp_rel <- taxa_barplot(physeq_16SASV, 
                       target_glom = "Genus", 
                       treatment_variable = "Host.genus", 
                       abundance_type = "relative", 
                       x_angle = 90, 
                       fill_variable = "Genus", 
                       facet_variable = "Diet", 
                       top_n_taxa = 20, 
                       palette = MG())

# Print the barplot for relative abundance
print(bp_rel$barplot)
} # }