Tidy a Phyloseq Object

This function cleans and tidies a phyloseq object by performing the following steps: - Fixes taxa names by removing any characters followed by '__' and any spaces after '__' - Sets taxonomic ranks to standard names and requires 7 ranks - Trims leading and trailing whitespace from taxa names - Replaces NA in the Phylum column with "Unidentified" - Removes taxa with zero counts - Removes taxa classified as "Chloroplast" at the Class level - Removes taxa classified as "Mitochondria" at the Family level

tidy_phyloseq(my_phyloseq)

Arguments

my_phyloseq

A phyloseq object containing the taxonomic and abundance data.

Value

A cleaned and tidied phyloseq object.

Examples

if (FALSE) { # \dontrun{
# Note: The DspikeIn package functions only work with datasets 
# that contain 7 taxonomic ranks (e.g., Kingdom, Phylum, Class,
# Order, Family, Genus, Species).

# Example usage:
#' # Tidy up the phyloseq object for easier downstream analysis
spiked_16S <- tidy_phyloseq(spiked_16S)

# The tidied phyloseq object can now be used in further analyses
} # }